requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.5.0
python-dp==1.1.1